backend/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
opencv-python==4.7.0.72
clip==1.0
numpy==1.24.3
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.1.0
pillow==9.5.0
